small refactoring for raycasting functions🧌
